+++ tcp.h
@@ -16,16 +16,18 @@
 extern "C" {
 #endif
 
+#include <stdint.h>
+
 /* Open a listen port on <host> and <port> and return the corresponding
  * file descriptor. If <host> is NULL the socket will listen on all
  * interfaces. If <port> is equal to 0, the socket will be bound to a
  * random local port (use tcpLocalPort() on the returned fd to find out
  * which). */
-int tcpListen(const char *host, int port);
+int tcpListen(const char *host, uint16_t port);
 
 /* Make a connection to <port> on <host> and return the corresponding
  * file descriptor. */
-int tcpConnect(const char *host, int port);
+int tcpConnect(const char *host, uint16_t port);
 
 /* Accept an incoming connection request on a listen socket */
 int tcpAccept(int sd);
